chore: clean the new filter UI experiment
