+++ frontend/src/scenes/notebooks/Nodes/NotebookNodePlaylist.tsx
@@ -15,7 +15,7 @@ import { SessionRecordingsPlaylist } from 'scenes/session-recordings/playlist/Se
 import { sessionRecordingPlayerLogic } from 'scenes/session-recordings/player/sessionRecordingPlayerLogic'
 import { IconComment } from 'lib/lemon-ui/icons'
 import { sessionRecordingPlayerLogicType } from 'scenes/session-recordings/player/sessionRecordingPlayerLogicType'
-import { RecordingsUniversalFilters } from 'scenes/session-recordings/filters/RecordingsUniversalFilters'
+import { RecordingsUniversalFiltersEmbed } from 'scenes/session-recordings/filters/RecordingsUniversalFiltersEmbed'
 import { PostHogErrorBoundary } from 'posthog-js/react'
 
 const Component = ({
@@ -121,7 +121,7 @@ export const Settings = ({
 
     return (
         <PostHogErrorBoundary>
-            <RecordingsUniversalFilters filters={filters} setFilters={setFilters} />
+            <RecordingsUniversalFiltersEmbed filters={filters} setFilters={setFilters} />
         </PostHogErrorBoundary>
     )
 }
