+++ frontend/src/scenes/session-recordings/playlist/SessionRecordingsPlaylist.tsx
@@ -2,16 +2,13 @@ import { LemonBadge, LemonButton, Link, Spinner } from '@posthog/lemon-ui'
 import { BindLogic, useActions, useValues } from 'kea'
 import { EmptyMessage } from 'lib/components/EmptyMessage/EmptyMessage'
 import { PropertyKeyInfo } from 'lib/components/PropertyKeyInfo'
-import { FEATURE_FLAGS } from 'lib/constants'
 import { LemonBanner } from 'lib/lemon-ui/LemonBanner'
-import { featureFlagLogic } from 'lib/logic/featureFlagLogic'
 import { useNotebookNode } from 'scenes/notebooks/Nodes/NotebookNodeContext'
 import { Playlist, PlaylistSection } from 'scenes/session-recordings/playlist/Playlist'
 import { urls } from 'scenes/urls'
 
 import { ReplayTabs } from '~/types'
 
-import { RecordingsUniversalFilters } from '../filters/RecordingsUniversalFilters'
 import {
     RecordingsUniversalFiltersEmbed,
     RecordingsUniversalFiltersEmbedButton,
@@ -56,8 +53,6 @@ export function SessionRecordingsPlaylist({
         totalFiltersCount,
     } = useValues(playlistLogic)
     const { maybeLoadSessionRecordings, setSelectedRecordingId, setFilters, resetFilters } = useActions(playlistLogic)
-
-    const { featureFlags } = useValues(featureFlagLogic)
 
     const notebookNode = useNotebookNode()
 
@@ -122,22 +117,11 @@ export function SessionRecordingsPlaylist({
                     sections={sections}
                     headerActions={<SessionRecordingsPlaylistTopSettings filters={filters} setFilters={setFilters} />}
                     filterActions={
-                        notebookNode || (!canMixFiltersAndPinned && !!logicProps.logicKey) ? null : featureFlags[
-                              FEATURE_FLAGS.REPLAY_FILTERS_IN_PLAYLIST
-                          ] === 'new' ? (
+                        notebookNode || (!canMixFiltersAndPinned && !!logicProps.logicKey) ? null : (
                             <RecordingsUniversalFiltersEmbedButton
                                 filters={filters}
                                 setFilters={setFilters}
                                 totalFiltersCount={totalFiltersCount}
-                            />
-                        ) : (
-                            <RecordingsUniversalFilters
-                                resetFilters={resetFilters}
-                                filters={filters}
-                                setFilters={setFilters}
-                                totalFiltersCount={totalFiltersCount}
-                                allowReplayHogQLFilters={allowHogQLFilters}
-                                allowReplayGroupsFilters={allowReplayGroupsFilters}
                             />
                         )
                     }
@@ -184,16 +168,14 @@ export function SessionRecordingsPlaylist({
                         )
                     }
                     filterContent={
-                        featureFlags[FEATURE_FLAGS.REPLAY_FILTERS_IN_PLAYLIST] === 'new' ? (
-                            <RecordingsUniversalFiltersEmbed
-                                resetFilters={resetFilters}
-                                filters={filters}
-                                setFilters={setFilters}
-                                totalFiltersCount={totalFiltersCount}
-                                allowReplayHogQLFilters={allowHogQLFilters}
-                                allowReplayGroupsFilters={allowReplayGroupsFilters}
-                            />
-                        ) : null
+                        <RecordingsUniversalFiltersEmbed
+                            resetFilters={resetFilters}
+                            filters={filters}
+                            setFilters={setFilters}
+                            totalFiltersCount={totalFiltersCount}
+                            allowReplayHogQLFilters={allowHogQLFilters}
+                            allowReplayGroupsFilters={allowReplayGroupsFilters}
+                        />
                     }
                 />
             </div>
